+++ src/renderer/reducers.ts
@@ -294,7 +294,13 @@ const controlReducer = (state: ControlState = INITIAL_CONTROL_STATE, action: Act
             };
         }
         case actions.UPDATE_MOUSE_IDLE_STATE: {
-            return {...state, ...action.payload}
+            return {...state, ...action.payload};
+        }
+        case actions.OPEN_MESSAGE_BOX: {
+            return {...state, messageBox: {...state.messageBox, ...action.payload, isOpen: true}};
+        }
+        case actions.CLOSE_MESSAGE_BOX: {
+            return {...state, messageBox: {...state.messageBox, ...action.payload, isOpen: false}};
         }
         default: {
             const newViews = viewsReducer(state.views, action, state.activeViewId);