+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Changes in v2.0.0-dev.4
 
+* The UI uses now web versions of native (electron) message boxes. (#47)
 * Is now using the Cate Web API to handle user preferences
 * Replaced CCI icons by simple round icons using GCOS ECV groupings as background color
 * Adapted UI fonds to conform to ESA style guidelines